Fix sprite-zero hit detection issue
The flag to indicate sprite zero is present on a scanline was being set properly, but then unset if any other sprites were present on the same scanline.  This caused crashes in Super Mario Bros if any other sprite moved high enough on the screen to interfere with the status bar.